requirements.txt
streamlit
pandas
pypdf
google-generativeai
gspread
google-auth
